we can now cleanly decouple api and no-api firecracker flavors
