+++ src/firecracker/src/main.rs
@@ -13,6 +13,8 @@ extern crate mmds;
 extern crate seccomp;
 extern crate vmm;
 
+mod api_server_adapter;
+
 use backtrace::Backtrace;
 use clap::{App, Arg};
 
@@ -21,20 +23,13 @@ use std::io;
 use std::panic;
 use std::path::PathBuf;
 use std::process;
-use std::sync::mpsc::{channel, Receiver, RecvError, Sender, TryRecvError};
-use std::sync::{Arc, RwLock};
-use std::thread;
 
-use api_server::{ApiServer, Error, VmmRequest, VmmResponse};
 use logger::{Metric, LOGGER, METRICS};
-use mmds::MMDS;
-use utils::eventfd::EventFd;
 use utils::terminal::Terminal;
 use utils::validators::validate_instance_id;
 use vmm::controller::VmmController;
 use vmm::resources::VmResources;
 use vmm::signal_handler::register_signal_handlers;
-use vmm::vmm_config;
 use vmm::vmm_config::instance_info::InstanceInfo;
 use vmm::EventLoopExitReason;
 
@@ -143,11 +138,6 @@ fn main() {
         )
         .get_matches();
 
-    let bind_path = cmd_arguments
-        .value_of("api_sock")
-        .map(PathBuf::from)
-        .expect("Missing argument: api_sock");
-
     // It's safe to unwrap here because clap's been provided with a default value
     let instance_id = cmd_arguments.value_of("id").unwrap().to_string();
 
@@ -159,132 +149,88 @@ fn main() {
         .parse::<u32>()
         .unwrap();
 
-    let start_time_us = cmd_arguments.value_of("start-time-us").map(|s| {
-        s.parse::<u64>()
-            .expect("'start-time-us' parameter expected to be of 'u64' type.")
-    });
-
-    let start_time_cpu_us = cmd_arguments.value_of("start-time-cpu-us").map(|s| {
-        s.parse::<u64>()
-            .expect("'start-time-cpu_us' parameter expected to be of 'u64' type.")
-    });
-
     let vmm_config_json = cmd_arguments
         .value_of("config-file")
         .map(fs::read_to_string)
         .map(|x| x.expect("Unable to open or read from the configuration file"));
 
-    let no_api = cmd_arguments.is_present("no-api");
+    let api_enabled = !cmd_arguments.is_present("no-api");
 
     LOGGER.set_instance_id(instance_id.clone());
-    let api_shared_info = Arc::new(RwLock::new(InstanceInfo {
-        id: instance_id,
-        started: false,
-        vmm_version: crate_version!().to_string(),
-    }));
 
-    let request_event_fd = EventFd::new(libc::EFD_NONBLOCK)
-        .map_err(Error::Eventfd)
-        .expect("Cannot create API Eventfd.");
-    let (to_vmm, from_api) = channel();
-    let (to_api, from_vmm) = channel();
+    match api_enabled {
+        false => run_without_api(seccomp_level, vmm_config_json),
+        true => {
+            let bind_path = cmd_arguments
+                .value_of("api_sock")
+                .map(PathBuf::from)
+                .expect("Missing argument: api_sock");
 
-    // Api enabled.
-    if !no_api {
-        // MMDS only supported with API.
-        let mmds_info = MMDS.clone();
-        let vmm_shared_info = api_shared_info.clone();
-        let to_vmm_event_fd = request_event_fd.try_clone().unwrap();
+            let start_time_us = cmd_arguments.value_of("start-time-us").map(|s| {
+                s.parse::<u64>()
+                    .expect("'start-time-us' parameter expected to be of 'u64' type.")
+            });
 
-        thread::Builder::new()
-            .name("fc_api".to_owned())
-            .spawn(move || {
-                match ApiServer::new(
-                    mmds_info,
-                    vmm_shared_info,
-                    to_vmm,
-                    from_vmm,
-                    to_vmm_event_fd,
-                )
-                .expect("Cannot create API server")
-                .bind_and_run(
-                    bind_path,
-                    start_time_us,
-                    start_time_cpu_us,
-                    seccomp_level,
-                ) {
-                    Ok(_) => (),
-                    Err(Error::Io(inner)) => match inner.kind() {
-                        io::ErrorKind::AddrInUse => {
-                            panic!("Failed to open the API socket: {:?}", Error::Io(inner))
-                        }
-                        _ => panic!(
-                            "Failed to communicate with the API socket: {:?}",
-                            Error::Io(inner)
-                        ),
-                    },
-                    Err(eventfd_err @ Error::Eventfd(_)) => {
-                        panic!("Failed to open the API socket: {:?}", eventfd_err)
-                    }
-                }
-            })
-            .expect("API thread spawn failed.");
-    }
-
-    run(
-        api_shared_info,
-        request_event_fd,
-        from_api,
-        to_api,
-        seccomp_level,
-        vmm_config_json,
-    );
+            let start_time_cpu_us = cmd_arguments.value_of("start-time-cpu-us").map(|s| {
+                s.parse::<u64>()
+                    .expect("'start-time-cpu_us' parameter expected to be of 'u64' type.")
+            });
+            let instance_info = InstanceInfo {
+                id: instance_id,
+                started: false,
+                vmm_version: crate_version!().to_string(),
+            };
+            api_server_adapter::run_with_api(
+                seccomp_level,
+                vmm_config_json,
+                bind_path,
+                instance_info,
+                start_time_us,
+                start_time_cpu_us,
+            )
+        }
+    };
 }
 
-/// Creates, starts then controls a vmm.
-///
-/// # Arguments
-///
-/// * `api_shared_info` - A parameter for storing information on the VMM (e.g the current state).
-/// * `api_event_fd` - An event fd used for receiving API associated events.
-/// * `from_api` - The receiver end point of the communication channel.
-/// * `seccomp_level` - The level of seccomp filtering used. Filters are loaded before executing
-///                     guest code. Can be one of 0 (seccomp disabled), 1 (filter by syscall
-///                     number) or 2 (filter by syscall number and argument values).
-/// * `config_json` - Optional parameter that can be used to configure the guest machine without
-///                   using the API socket.
-fn run(
-    api_shared_info: Arc<RwLock<InstanceInfo>>,
-    api_event_fd: EventFd,
-    from_api: Receiver<VmmRequest>,
-    to_api: Sender<VmmResponse>,
+// Configure and start a microVM as described by the command-line JSON.
+fn build_microvm_from_json(
     seccomp_level: u32,
-    config_json: Option<String>,
-) {
+    config_json: String,
+    epoll_context: &mut vmm::EpollContext,
+) -> (VmResources, vmm::Vmm) {
+    let vm_resources = VmResources::from_json(&config_json, crate_version!().to_string())
+        .unwrap_or_else(|err| {
+            error!(
+                "Configuration for VMM from one single json failed: {:?}",
+                err
+            );
+            process::exit(i32::from(vmm::FC_EXIT_CODE_BAD_CONFIGURATION));
+        });
+    let vmm = vmm::builder::build_microvm(&vm_resources, epoll_context, seccomp_level)
+        .unwrap_or_else(|err| {
+            error!(
+                "Building VMM configured from cmdline json failed: {:?}",
+                err
+            );
+            process::exit(i32::from(vmm::FC_EXIT_CODE_BAD_CONFIGURATION));
+        });
+    info!("Successfully started microvm that was configured from one single json");
+
+    (vm_resources, vmm)
+}
+
+fn run_without_api(seccomp_level: u32, config_json: Option<String>) {
     // The driving epoll engine.
     let mut epoll_context = vmm::EpollContext::new().expect("Cannot create the epoll context.");
-    epoll_context
-        .add_epollin_event(&api_event_fd, vmm::EpollDispatch::VmmActionRequest)
-        .expect("Cannot add vmm control_fd to epoll.");
 
-    // These will be populated from JSON or API.
-    let mut vm_resources = VmResources::default();
-
-    // Build and start the microVM.
-    let vmm = configure_and_build_microvm(
-        &api_event_fd,
-        &from_api,
-        &to_api,
+    let (vm_resources, vmm) = build_microvm_from_json(
         seccomp_level,
-        config_json,
+        // Safe to unwrap since no-api requires this to be set.
+        config_json.unwrap(),
         &mut epoll_context,
-        &mut vm_resources,
     );
 
-    api_shared_info.write().unwrap().started = true;
-
     let mut vm_controller: VmmController = VmmController::new(epoll_context, vm_resources, vmm);
-
     let exit_code = loop {
         match vm_controller.run_event_loop() {
             Err(e) => {
@@ -297,238 +243,10 @@ fn run(
                     break vmm::FC_EXIT_CODE_OK;
                 }
                 EventLoopExitReason::ControlAction => {
-                    if let Err(exit_code) =
-                        vmm_control_event(&mut vm_controller, &api_event_fd, &from_api, &to_api)
-                    {
-                        break exit_code;
-                    }
+                    warn!("Spurious control event");
                 }
             },
         };
     };
-
     vm_controller.stop(i32::from(exit_code));
-}
-
-/// Builds and starts a microVM. It either uses a config json or receives API requests to
-/// get the microVM configuration.
-///
-/// Returns a running `Vmm` object.
-fn configure_and_build_microvm(
-    api_event_fd: &EventFd,
-    from_api: &Receiver<VmmRequest>,
-    to_api: &Sender<VmmResponse>,
-    seccomp_level: u32,
-    config_json: Option<String>,
-    // FIXME: epoll context can be polluted by failing boot attempts
-    epoll_context: &mut vmm::EpollContext,
-    vm_resources: &mut VmResources,
-) -> vmm::Vmm {
-    use vmm::{ErrorKind, VmmActionError};
-
-    // Start a microVm configured from command-line JSON.
-    if let Some(ref json) = config_json {
-        *vm_resources =
-            VmResources::from_json(json, crate_version!().to_string()).unwrap_or_else(|err| {
-                error!(
-                    "Setting configuration for VMM from one single json failed: {:?}",
-                    err
-                );
-                process::exit(i32::from(vmm::FC_EXIT_CODE_BAD_CONFIGURATION));
-            });
-        match vmm::builder::build_microvm(&vm_resources, epoll_context, seccomp_level) {
-            Ok(vmm) => {
-                info!("Successfully started microvm that was configured from one single json");
-                return vmm;
-            }
-            Err(err) => {
-                error!(
-                    "Building VMM configured from cmdline json failed: {:?}",
-                    err
-                );
-                process::exit(i32::from(vmm::FC_EXIT_CODE_BAD_CONFIGURATION));
-            }
-        };
-    }
-
-    // Configure and start microVM through successive API calls.
-    // Iterate through API calls to configure microVm.
-    // The loop breaks when a microVM is successfully started, and returns a running Vmm.
-    loop {
-        let mut built_vmm = None;
-        match from_api.recv() {
-            Ok(vmm_request) => {
-                use api_server::VmmAction::*;
-                let action_request = *vmm_request;
-
-                // Also consume the API event. This is safe since communication
-                // between this thread and the API thread is synchronous.
-                let _ = api_event_fd.read().map_err(|e| {
-                    error!("VMM: Failed to read the API event_fd: {}", e);
-                    process::exit(i32::from(vmm::FC_EXIT_CODE_GENERIC_ERROR));
-                });
-
-                let response = match action_request {
-                    /////////////////////////////////////////
-                    // Supported operations allowed pre-boot.
-                    ConfigureBootSource(boot_source_body) => vm_resources
-                        .set_boot_source(boot_source_body)
-                        .map(|_| api_server::VmmData::Empty)
-                        .map_err(|e| VmmActionError::BootSource(ErrorKind::User, e)),
-                    ConfigureLogger(logger_description) => vmm_config::logger::init_logger(
-                        logger_description,
-                        crate_version!().to_string(),
-                    )
-                    .map(|_| api_server::VmmData::Empty)
-                    .map_err(|e| VmmActionError::Logger(ErrorKind::User, e)),
-                    GetVmConfiguration => Ok(api_server::VmmData::MachineConfiguration(
-                        vm_resources.vm_config().clone(),
-                    )),
-                    InsertBlockDevice(block_device_config) => vm_resources
-                        .set_block_device(block_device_config)
-                        .map(|_| api_server::VmmData::Empty)
-                        .map_err(|e| e.into()),
-                    InsertNetworkDevice(netif_body) => vm_resources
-                        .set_net_device(netif_body)
-                        .map(|_| api_server::VmmData::Empty)
-                        .map_err(|e| e.into()),
-                    SetVsockDevice(vsock_cfg) => {
-                        vm_resources.set_vsock_device(vsock_cfg);
-                        Ok(api_server::VmmData::Empty)
-                    }
-                    SetVmConfiguration(machine_config_body) => vm_resources
-                        .set_vm_config(machine_config_body)
-                        .map(|_| api_server::VmmData::Empty)
-                        .map_err(|e| e.into()),
-                    UpdateBlockDevicePath(drive_id, path_on_host) => vm_resources
-                        .update_block_device_path(drive_id, path_on_host)
-                        .map(|_| api_server::VmmData::Empty)
-                        .map_err(|e| e.into()),
-                    UpdateNetworkInterface(netif_update) => vm_resources
-                        .update_net_rate_limiters(netif_update)
-                        .map(|_| api_server::VmmData::Empty)
-                        .map_err(|e| e.into()),
-                    StartMicroVm => {
-                        vmm::builder::build_microvm(&vm_resources, epoll_context, seccomp_level)
-                            .map(|vmm| {
-                                built_vmm = Some(vmm);
-                                api_server::VmmData::Empty
-                            })
-                    }
-
-                    ///////////////////////////////////
-                    // Operations not allowed pre-boot.
-                    FlushMetrics => Err(VmmActionError::Logger(
-                        ErrorKind::User,
-                        vmm_config::logger::LoggerConfigError::FlushMetrics(
-                            "Cannot flush metrics before starting microVM.".to_string(),
-                        ),
-                    )),
-                    RescanBlockDevice(_) => {
-                        Err(vmm_config::drive::DriveError::OperationNotAllowedPreBoot.into())
-                    }
-                    SendCtrlAltDel => Err(vmm::error::VmmActionError::OperationNotSupportedPreBoot),
-                };
-                // Send back the result.
-                to_api
-                    .send(Box::new(response))
-                    .map_err(|_| ())
-                    .expect("one-shot channel closed");
-                // If microVM was successfully started, return the Vmm.
-                if let Some(vmm) = built_vmm {
-                    break vmm;
-                }
-            }
-            Err(RecvError) => {
-                panic!("The channel's sending half was disconnected. Cannot receive data.");
-            }
-        };
-    }
-}
-
-/// Handles the control event.
-/// Receives and runs the Vmm action and sends back a response.
-/// Provides program exit codes on errors.
-fn vmm_control_event(
-    controller: &mut VmmController,
-    api_event_fd: &EventFd,
-    from_api: &Receiver<VmmRequest>,
-    to_api: &Sender<VmmResponse>,
-) -> Result<(), u8> {
-    use vmm::{ErrorKind, VmmActionError};
-
-    api_event_fd.read().map_err(|e| {
-        error!("VMM: Failed to read the API event_fd: {}", e);
-        vmm::FC_EXIT_CODE_GENERIC_ERROR
-    })?;
-
-    match from_api.try_recv() {
-        Ok(vmm_request) => {
-            use api_server::VmmAction::*;
-            let action_request = *vmm_request;
-            let response = match action_request {
-                ///////////////////////////////////
-                // Supported operations allowed post-boot.
-                FlushMetrics => controller
-                    .flush_metrics()
-                    .map(|_| api_server::VmmData::Empty),
-                GetVmConfiguration => Ok(api_server::VmmData::MachineConfiguration(
-                    controller.vm_config().clone(),
-                )),
-                RescanBlockDevice(drive_id) => controller
-                    .rescan_block_device(&drive_id)
-                    .map(|_| api_server::VmmData::Empty),
-                #[cfg(target_arch = "x86_64")]
-                SendCtrlAltDel => controller
-                    .send_ctrl_alt_del()
-                    .map(|_| api_server::VmmData::Empty),
-                UpdateBlockDevicePath(drive_id, path_on_host) => controller
-                    .update_block_device_path(drive_id, path_on_host)
-                    .map(|_| api_server::VmmData::Empty),
-                UpdateNetworkInterface(netif_update) => controller
-                    .update_net_rate_limiters(netif_update)
-                    .map(|_| api_server::VmmData::Empty),
-
-                ///////////////////////////////////
-                // Operations not allowed post-boot.
-                ConfigureBootSource(_) => Err(VmmActionError::BootSource(
-                    ErrorKind::User,
-                    vmm_config::boot_source::BootSourceConfigError::UpdateNotAllowedPostBoot,
-                )),
-                ConfigureLogger(_) => Err(VmmActionError::Logger(
-                    ErrorKind::User,
-                    vmm_config::logger::LoggerConfigError::InitializationFailure(
-                        "Cannot initialize logger after boot.".to_string(),
-                    ),
-                )),
-                InsertBlockDevice(_) => {
-                    Err(vmm_config::drive::DriveError::UpdateNotAllowedPostBoot.into())
-                }
-                InsertNetworkDevice(_) => {
-                    Err(vmm_config::net::NetworkInterfaceError::UpdateNotAllowedPostBoot.into())
-                }
-                SetVsockDevice(_) => Err(VmmActionError::VsockConfig(
-                    ErrorKind::User,
-                    vmm_config::vsock::VsockError::UpdateNotAllowedPostBoot,
-                )),
-                SetVmConfiguration(_) => {
-                    Err(vmm_config::machine_config::VmConfigError::UpdateNotAllowedPostBoot.into())
-                }
-
-                StartMicroVm => Err(vmm::error::StartMicrovmError::MicroVMAlreadyRunning.into()),
-            };
-            // Send back the result.
-            to_api
-                .send(Box::new(response))
-                .map_err(|_| ())
-                .expect("one-shot channel closed");
-        }
-        Err(TryRecvError::Empty) => {
-            warn!("Got a spurious notification from api thread");
-        }
-        Err(TryRecvError::Disconnected) => {
-            panic!("The channel's sending half was disconnected. Cannot receive data.");
-        }
-    };
-    Ok(())
 }
